\name{EMMREML}
\alias{EMMREML}
\alias{EMMREML}
\docType{package}
\title{
Fitting mixed models with known covariance structures.
}
\description{
The main functions areemmreml, and emmremlMultiKernel.emmreml solves a mixed model with known covariance structure. emmremlMultiKernel is a wrapper for emmreml to handle multiple random components with known covariance structures.
}
\details{
\tabular{ll}{
Package: \tab EMMREML\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-02-24\cr
License: \tab GPL-2\cr
}
}
\author{
Deniz Akdemir

Maintainer: Deniz Akdemir <deniz.akdemir.work@gmail.com>
}
\references{
	Akdemir, Deniz. "Genomic Prediction of Quantitative Traits using Sparse and Locally Epistatic Models." arXiv preprint arXiv:1402.2026 (2014).
}


