\name{EMMREML}
\alias{EMMREML}
\alias{EMMREML}
\docType{package}
\title{
Fitting mixed models with known covariance structures.
}
\description{
The main functions are emmreml, and emmremlMultiKernel.emmreml solves a mixed model with known covariance structure using the EMMA algorithm in Kang et.al. (2008). emmremlMultiKernel is a wrapper for emmreml to handle multiple random components with known covariance structures. The function emmremlMultivariate solves a multivariate gaussian mixed model with known covariance structure using the ECM algorithm in Zhou and Stephens (2012).
}
\details{
\tabular{ll}{
Package: \tab EMMREML\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2014-05-23\cr
License: \tab GPL-2\cr
}
}
\author{
Deniz Akdemir, Okeke Uche Godfrey

Maintainer: Deniz Akdemir <deniz.akdemir.work@gmail.com>
}
\references{
	
Efficient control of population structure in model organism association mapping. Kang, Hyun Min and Zaitlen, Noah A and Wade, Claire M and Kirby, Andrew and Heckerman, David and Daly, Mark J and Eskin, Eleazar. Genetics, 2008.
  
  
Genome-wide efficient mixed-model analysis for association studies. Zhou, Xiang and Stephens, Matthew. Nature genetics, 2012.
}


